+++ pybis/src/python/CHANGELOG.md
@@ -1,3 +1,22 @@
+## Changes with pybis-1.18.3
+
+- prevent other users to read the saved token (chmod 600)
+
+## Changes with pybis-1.18.2
+
+- added deletion to transaction
+
+## Changes with pybis-1.18.1
+
+- fixed del_parents() bug accidentally introduced in 1.18.0
+
+## Changes with pybis-1.18.0
+
+- speed improvement when searching for samples and dataSets and then cycling through the results
+- implemented search for number comparison, date comparison, string comparison (<, >, <=, >=)
+- implemented search for parents identities and properties
+- fixed minor bugs when connecting
+
 ## Changes with pybis-1.17.4
 
 - fixed another vocabularies update bug